♻️ refactor(layout): Refactor ssc layout
