+++ src/app/chat/(mobile)/mobile/layout.mobile.tsx
@@ -1,0 +1,9 @@
+import { PropsWithChildren } from 'react';
+
+import AppLayoutMobile from '@/layout/AppLayout.mobile';
+
+import Header from '../../components/ChatHeader/Mobile';
+
+export default ({ children }: PropsWithChildren) => (
+  <AppLayoutMobile navBar={<Header />}>{children}</AppLayoutMobile>
+);
