+++ src/app/chat/(mobile)/mobile/page.tsx
@@ -3,19 +3,17 @@
 import { memo } from 'react';
 import { Flexbox } from 'react-layout-kit';
 
-import AppMobileLayout from '@/layout/AppMobileLayout';
-
 import Conversation from '../../features/Conversation';
-import Header from '../features/ChatHeader';
 import ChatInput from '../features/ChatInput';
 import Topics from '../features/Topics';
+import Layout from './layout.mobile';
 
 const Chat = memo(() => (
-  <AppMobileLayout navBar={<Header />}>
+  <Layout>
     <Flexbox height={'calc(100vh - 44px)'} horizontal>
       <Conversation chatInput={<ChatInput />} mobile />
       <Topics />
     </Flexbox>
-  </AppMobileLayout>
+  </Layout>
 ));
 export default Chat;
