+++ src/app/chat/settings/(desktop)/index.tsx
@@ -1,0 +1,14 @@
+'use client';
+
+import { memo } from 'react';
+
+import EditPage from '../features/EditPage';
+import Layout from './layout.responsive';
+
+const ChatSettings = memo(() => (
+  <Layout>
+    <EditPage />
+  </Layout>
+));
+
+export default ChatSettings;
