+++ src/app/settings/(desktop)/features/Header.tsx
@@ -1,5 +1,4 @@
-import { MobileNavBar, MobileNavBarTitle } from '@lobehub/ui';
-import { useRouter } from 'next/navigation';
+import { ChatHeader, ChatHeaderTitle } from '@lobehub/ui';
 import { memo } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -8,13 +7,14 @@ import { useGlobalStore } from '@/store/global';
 const Header = memo(() => {
   const { t } = useTranslation('setting');
   const tab = useGlobalStore((s) => s.settingsTab);
-  const router = useRouter();
 
   return (
-    <MobileNavBar
-      center={<MobileNavBarTitle title={t(`tab.${tab}`)} />}
-      onBackClick={() => router.back()}
-      showBackButton
+    <ChatHeader
+      left={
+        <div style={{ paddingLeft: 8 }}>
+          <ChatHeaderTitle title={t(`tab.${tab}`)} />
+        </div>
+      }
     />
   );
 });
