+++ src/app/settings/(desktop)/index.tsx
@@ -1,0 +1,1 @@
+export { default } from '../common/page';
