+++ src/app/settings/(mobile)/features/Header/Home.tsx
@@ -1,0 +1,8 @@
+import { Logo, MobileNavBar } from '@lobehub/ui';
+import { memo } from 'react';
+
+const Header = memo(() => {
+  return <MobileNavBar center={<Logo type={'text'} />} />;
+});
+
+export default Header;
