+++ src/app/settings/(mobile)/index.tsx
@@ -1,0 +1,1 @@
+export { default } from './mobile';
