+++ src/app/settings/(mobile)/mobile/page.tsx
