+++ src/app/settings/common/layout.tsx
@@ -1,0 +1,1 @@
+export { default } from '../layout.responsive';
