--- src/app/settings/mobile/page.tsx
@@ -1,7 +1,0 @@
-import Page from './index';
-
-const Index = () => {
-  return <Page />;
-};
-
-export default Index;
