+++ src/app/settings/page.tsx
@@ -1,1 +1,12 @@
-export { default } from './common/page';
+import { isMobileDevice } from '@/utils/responsive';
+
+import DesktopPage from './(desktop)';
+import MobilePage from './(mobile)';
+
+const Page = () => {
+  const mobile = isMobileDevice();
+
+  return mobile ? <MobilePage /> : <DesktopPage />;
+};
+
+export default Page;
