+++ src/app/welcome/MobileLayout.tsx
@@ -4,19 +4,19 @@ import { Logo, MobileNavBar } from '@lobehub/ui';
 import { useTheme } from 'antd-style';
 import { PropsWithChildren, memo } from 'react';
 
-import AppMobileLayout from '@/layout/AppMobileLayout';
+import AppLayoutMobile from '@/layout/AppLayout.mobile';
 
 const MobileLayout = memo<PropsWithChildren>(({ children }) => {
   const theme = useTheme();
 
   return (
-    <AppMobileLayout
+    <AppLayoutMobile
       navBar={<MobileNavBar center={<Logo type={'text'} />} />}
       showTabBar
       style={{ background: theme.colorBgLayout }}
     >
       {children}
-    </AppMobileLayout>
+    </AppLayoutMobile>
   );
 });
